flatpak: Make it possible to build from the non-master branch
An ugly shell script will create a json manifest file for Flatpak called
"org.gnome.FeedReader-BRANCH_NAME.json", adding the branch to build from
for FeedReader.

To make this easier, and because "master" is the default
branch in Flatpak manifests, remove the branch definition and replace
the branch by tacking it onto the end of the URL's line.

Example usage:
make BRANCH=stable