Make favicon loading in the UI a lot faster
We used to reload the entire favicon cache when we couldn't
find an icon, which meant that if you have 200 feeds, we would
reload the favicon cache 200 times in the initial startup.

This changes it so we only load the requested icons. In my testing,
FeedReader loads almost instantly, even with my crazy number of
feeds.

Part of #557