Change ownCloud ping to valid method
The previous method (a PUT to api/v1-2/) is not a part of the API description. Not only does The Arsse reject it, but NextCloud News itself gets confused and redirects to the user's NextCloud Files main page instead. It works, but may not in future NextCloud versions.

Instead, a GET to the /version route is both valid and low-overhead, and should continue to work in perpetuity.

I also corrected a repeated typo I noticed while I was looking around.