Fix article tagging
A commit a long time ago removed the tags column from articles and replaced it with
the taggings table, but it looks like the code to rename and delete tags, and also
to show all tagged articles was broken.

This fixes all of the references to the old tags column and switches them to
reference taggings instead. I also inlined the all-tags subquery.

This should fix #840