Auto-fill missing schema in TT-RSS URL
This should make it less confusing if users pull in a host name with no
scheme.

I made it default to https:// for safety but this should make it
sufficiently obvious if the right scheme is http:// (I intentionally
made it update the input box so people can see what we're doing).