Change the default font to use the system "document font"
Previously we defaulted to Droid fonts, but we shouldn't be trying to
make assumptions about what fonts users have installed or what they
want to use here.

This makes it so we use the system document font by default and we
still allow users to override it at-will in the settings.

See #877