Deduplicate feeds by feed URL, not link element.
Previously, we used the RSS <link> to determine if a feed was a duplicate, but
it's common for multiple feeds to exist for a single site. This changes the
logic to just deduplicate based on the actual link given to us by a user.

I also fixed a bug where we checked more than 1 duplicate instead of just
checking for any duplicates.

See #899 and #883