Re-escape HTML data after Htmlclean
This works by parsing HTML data into normal text. However, when we're
done with it, we want this to become valid HTML again. If we leave
things like & parsed, we can get situations where the title/content
of a page depends on how many times we run it through Htmlclean.

Fixes #918